Bassicly testing has green celll population
In the main code it pretty much works flawlessey. It removes everything after @  when compapring emaills and fills this years thing in about four  min. Only doesnt mark when email are differnt before the @. The testing is same but just popullates in green. Need to find a way that reduces the amount of requests it uses so takes shorter time.